+++ packages/react-conformance-make-styles/src/matchers/index.ts
@@ -1,0 +1,5 @@
+import { toContainClassNameLastInCalls } from './toContainClassNameLastInCalls';
+
+expect.extend({
+  toContainClassNameLastInCalls,
+});