+++ packages/react-menu/src/common/mockUseMenuContext.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { MenuContextValue, useMenuContext } from '../contexts/menuContext';
+import { useMenuContext } from '../contexts/menuContext';
+import type { MenuContextValue } from '../contexts/menuContext';
 
 /**
  * A test utility to mock the useMenuContext hook that relies on context selector