+++ packages/react-menu/src/selectable/useCheckmarkStyles.ts
@@ -1,6 +1,6 @@
 import { mergeClasses, makeStyles } from '@fluentui/react-make-styles';
-import { MenuItemSelectableState } from './types';
-import { MenuItemSlots } from '../components/MenuItem/MenuItem.types';
+import type { MenuItemSelectableState } from './types';
+import type { MenuItemSlots } from '../components/MenuItem/MenuItem.types';
 
 const useStyles = makeStyles({
   root: {