+++ packages/react-slider/src/components/Slider/Slider.tsx
@@ -1,12 +1,11 @@
 import * as React from 'react';
 import { useSlider } from './useSlider';
-import { SliderProps } from './Slider.types';
 import { renderSlider } from './renderSlider';
 import { useSliderStyles } from './useSliderStyles';
+import type { SliderProps } from './Slider.types';
 
 /**
- * Define a styled Slider, using the `useSlider` hook.
- * {@docCategory Slider}
+ * Define a styled Slider, using the `useSlider` hook
  */
 export const Slider = React.forwardRef<HTMLElement, SliderProps>((props, ref) => {
   const state = useSlider(props, ref);