+++ packages/common-styles/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/common-styles
 
-This log was last generated on Thu, 19 Aug 2021 07:41:35 GMT and should not be manually modified.
+This log was last generated on Wed, 25 Aug 2021 07:35:19 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.16](https://github.com/microsoft/fluentui/tree/@fluentui/common-styles_v1.0.16)
+
+Wed, 25 Aug 2021 07:35:19 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/common-styles_v1.0.15..@fluentui/common-styles_v1.0.16)
+
+### Patches
+
+- Bump @fluentui/style-utilities to v8.3.1 ([PR #19481](https://github.com/microsoft/fluentui/pull/19481) by arujain@microsoft.com)
 
 ## [1.0.15](https://github.com/microsoft/fluentui/tree/@fluentui/common-styles_v1.0.15)